+++ server/example.env
@@ -1,5 +1,6 @@
 API_KEYS=your-api-key-1,your-api-key-2,your-api-key-3
 LINE_MAX_AUDIO_FILES=
+LINE_AUDIO_FILES_DIR=
 
 AGENT_1_NAME=
 AGENT_1_LLM_BASE_URL=
